verify-from-clone.sh: доказывает, что репозиторий самодостаточен
Утверждение сильнее прежнего: не «наше дерево работает», а
git clone -> скрипты ИЗ репозитория -> сборка -> тесты = рабочий multi-region
Valhalla, без единого файла с машины разработчика.

Прогон на чистом клоне be4fc3e, все восемь пунктов OK:
  дерево откатывается к стоку -> 0 маркеров
  apply-patches.sh применяется, второй прогон ничего не меняет
  сборка: 0 ошибок
  собираются все пять инструментов
  gid_test: инвариант GraphId зелёный
  регрессия: 109.300 / 3.684 / 4.112 км совпадают со стоком upstream 3.6.3
  матрица: 7 маршрутов, 2 NO ROUTE, 0 потерь региона, 0 разрывов геометрии

Попутно два дефекта СТЕНДА, каждый изображал дефект продукта:

  final.sh передавал координаты теста 1 как "47.2075 27.8000" -- в кавычках
  это ОДИН аргумент, программа падала на usage, и тест молча показывал
  пустоту. Я раньше списал это на буферизацию вывода; это было неверно.

  verify считал NO ROUTE по всей строке и ловил заголовки вида «обязан быть
  NO ROUTE», получая 4 вместо 2.

В gid_test снято предупреждение об неоднозначности `tagged + 1`: явный
uint64_t(1) заодно фиксирует, какой operator+ проверяется. Предупреждений там
было больше всех, и из-за `g++ | head -3` компилятор получал SIGPIPE -- тест
выглядел несобирающимся, хотя дело было в пайпе.

+++ wedrive/tools/gid_test.cc
@@ -43,7 +43,7 @@ int main() {
   check("set_id(11) сохраняет region", a.region() == kRegion && a.id() == 11);
   check("set_id(11) не трогает tileid/level", a.tileid() == kTile && a.level() == kLevel);
 
-  const GraphId b = tagged + 1;
+  const GraphId b = tagged + uint64_t(1);
   check("operator+(1) сохраняет region", b.region() == kRegion && b.id() == kId + 1);
 
   GraphId c = tagged;
@@ -73,7 +73,7 @@ int main() {
   check("set_id идентичен stock-конструкции", p1.value == s1.value);
 
   check("operator+ идентичен stock-конструкции",
-        (plain + 1).value == GraphId(kTile, kLevel, kId + 1).value);
+        (plain + uint64_t(1)).value == GraphId(kTile, kLevel, kId + 1).value);
 
   GraphId p2 = plain;
   ++p2;
